new: o2m field support.
Some code was cleaned and factorized with Many2Many implementation. We could'nt
use the current ``instance.web.FieldOne2Many`` as it's using a very different
approach than ``instance.web.Many2Many*`` that was not cleaned enough to be
re-usable.

It felt that o2m fields were not that different from m2m, and this proved to be
right and led to many factorizations, and a clearer view IFAIC of different
classe's responsability.

Some bug where spotted, and should be fixed in the next commits.